# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MyPetly API Documentation",
        "description": "This is the API documentation for MyPetly.",
        "contact": {
            "email": "support@mypetly.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/animals": {
            "get": {
                "tags": [
                    "Animals"
                ],
                "summary": "Get all animals",
                "description": "Retrieve a list of all animals for the authenticated user",
                "operationId": "9a489950fd27cd9c56fd59b937d0f501",
                "responses": {
                    "200": {
                        "description": "List of animals",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Buddy"
                                            },
                                            "species": {
                                                "type": "string",
                                                "example": "Dog"
                                            },
                                            "breed": {
                                                "type": "string",
                                                "example": "Labrador"
                                            },
                                            "birthdate": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2020-05-10"
                                            },
                                            "photo_url": {
                                                "type": "string",
                                                "example": "http://example.com/storage/animals/1.jpg"
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "active"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "An error occurred."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Detailed error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Animals"
                ],
                "summary": "Create a new animal",
                "description": "Add a new animal profile for the authenticated user",
                "operationId": "34e7a0c362b9e5cd0abcd38155cc2c89",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Buddy"
                                    },
                                    "species": {
                                        "type": "string",
                                        "example": "Dog"
                                    },
                                    "breed": {
                                        "type": "string",
                                        "example": "Labrador"
                                    },
                                    "birthdate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2020-05-10"
                                    },
                                    "photo": {
                                        "description": "Animal photo file",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Animal created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Buddy"
                                        },
                                        "species": {
                                            "type": "string",
                                            "example": "Dog"
                                        },
                                        "breed": {
                                            "type": "string",
                                            "example": "Labrador"
                                        },
                                        "birthdate": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2020-05-10"
                                        },
                                        "photo_url": {
                                            "type": "string",
                                            "example": "http://example.com/storage/animals/1.jpg"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "active"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "name": [
                                                    "The name field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/animals/{id}/collar": {
            "post": {
                "tags": [
                    "Animals"
                ],
                "summary": "Attach a collar to an animal",
                "description": "Attach a unique NFC/QR collar to the specified animal",
                "operationId": "4540af4dc7cead5a284bd4f2e762b74e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the animal",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "collar_id"
                                ],
                                "properties": {
                                    "collar_id": {
                                        "type": "string",
                                        "example": "12345ABC"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Collar attached successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Collar attached successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Animal not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Animal not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "collar_id": [
                                                    "The collar ID must be unique."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/animals/{id}/lost": {
            "put": {
                "tags": [
                    "Animals"
                ],
                "summary": "Mark an animal as lost",
                "description": "Update the status of an animal to 'lost'",
                "operationId": "fce200f61ecfd59ec575bedbed7d815f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the animal",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Animal marked as lost",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Animal marked as lost."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Animal not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Animal not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}